change to single activity using navigation
