+++ app/build.gradle
@@ -15,9 +15,19 @@ android {
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    kotlinOptions {
+        jvmTarget = JavaVersion.VERSION_1_8.toString()
+    }
+
     buildTypes {
 
-        debug{
+        debug {
             buildConfigField "String", "BASE_URL", "$github_url"
             buildConfigField "String", "REPOSITORY_TAG", "$repository_tag"
         }
@@ -35,11 +45,15 @@ dependencies {
 
     // Default
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "androidx.constraintlayout:constraintlayout:$constraintlayout"
     implementation "androidx.appcompat:appcompat:$appcompat"
     implementation "androidx.core:core-ktx:$android_ktx"
     implementation "androidx.recyclerview:recyclerview:$reclycerview"
+
+    // Navigation
+    implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
+    implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
 
     // Retrofit
     implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
@@ -49,7 +63,11 @@ dependencies {
     implementation "com.squareup.picasso:picasso:$picasso_version"
 
     // Test
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     testImplementation "junit:junit:$jnuit_jnuit"
     androidTestImplementation "androidx.test.ext:junit:$androidx_jnuit"
     androidTestImplementation "androidx.test.espresso:espresso-core:$espresso_core"
+
 }
+
+apply plugin: "androidx.navigation.safeargs"
