+++ app/src/main/java/com/concrete/desafio_android/ui/activity/HostActivity.kt
@@ -1,0 +1,14 @@
+package com.concrete.desafio_android.ui.activity
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.concrete.desafio_android.R
+
+class HostActivity: AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_host)
+    }
+
+}
