+++ app/src/main/res/layout/fragment_pull_request.xml
@@ -2,8 +2,8 @@
 <androidx.recyclerview.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/list_java_repositories"
+    android:id="@+id/pull_request_list"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-    tools:listitem="@layout/list_item_repository" />
+    tools:listitem="@layout/list_item_pull_request" />
