+++ app/src/main/res/values/strings.xml
@@ -2,19 +2,21 @@
     <string name="app_name">Desafio Android</string>
 
     <!--  begin region repositories  -->
-    <string name="home_activity_label">Github Java Top Repositories</string>
+    <string name="repositories_fragment_label">Github Java Top Repositories</string>
     <string name="repository_avatar_content_description">Repository owner avatar image.</string>
     <string name="star_icon_content_description">Repository star count icon.</string>
     <string name="fork_icon_content_description">Repository fork count icon.</string>
     <!--  end region repositories  -->
 
     <!--  begin region pullRequest  -->
-    <string name="repository_not_loaded_error">Github Java Top Repositories</string>
     <string name="pull_request_avatar_content_description">Pull request creator avatar image.</string>
     <!--  end region pullRequest  -->
 
     <!--  begin region error  -->
     <string name="connection_problem">Could not connect with github.</string>
     <string name="unknown_error">An unexpected error occurred.</string>
-    <!--  end region error  -->
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <!--  begin region error  -->
 </resources>
