+++ build.gradle
@@ -19,6 +19,7 @@ buildscript {
         espresso_core = '3.1.1'
         retrofit_version = '2.6.0'
         picasso_version = '2.71828'
+        nav_version = '2.2.1'
         github_url = '"https://api.github.com"'
         repository_tag = '"repository"'
     }
@@ -26,6 +27,7 @@ buildscript {
     dependencies {
         classpath "com.android.tools.build:gradle:$gradle_verison"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
